still working on delete button

+++ src/components/MyPostsCard.jsx
@@ -2,10 +2,13 @@ import { deletePost } from "../API";
 
 export default function MyPostsCard({ post, token }) {
     //TODO show messages once send message is functional
-    const { _id, title, author, description, location, price, willDeliver, messages, active} = post;
+    const { _id, title, description, location, price, willDeliver, messages, active, isAuthor} = post;
     // const [showForm, setShowForm] = useState(false)
     async function handleClick(_id, token) {
-        await deletePost(_id, token);
+        if (isAuthor) {
+            await deletePost(_id, token);
+        }
+        
     }
     return (
       <div className="myPostsCard" key={_id}>
